refactor: delete bossSelection.ts; relocate helpers module-private into muleBossSlate.ts
Slice 4/4 of the MuleBossSlate migration (PRD #170). All selection-key
helpers (makeKey, parseKey, toggleBoss, validateBossSelection,
countWeeklySelections, getFamilies, hardestDifficulty, TIER_ORDER, etc.)
become module-private inside src/data/muleBossSlate.ts. Remaining callers
(bossPresets, income, useMules, BossMatrix) use MuleBossSlate.from(...) or
inline the minimal key pattern they need.

Closes #174
