+++ src/components/__tests__/MuleCharacterCard.test.tsx
@@ -5,10 +5,9 @@ import { SortableContext, rectSortingStrategy } from '@dnd-kit/sortable'
 import { MuleCharacterCard, MuleCharacterCardOverlay } from '../MuleCharacterCard'
 import type { Mule } from '../../types'
 import { bosses } from '../../data/bosses'
-import { makeKey } from '../../data/bossSelection'
 
 const LUCID = bosses.find((b) => b.family === 'lucid')!.id
-const HARD_LUCID = makeKey(LUCID, 'hard', 'weekly')
+const HARD_LUCID = `${LUCID}:hard:weekly`
 
 const baseMule: Mule = {
   id: 'test-mule-1',
