+++ src/data/bossPresets.ts
@@ -1,6 +1,5 @@
-import type { BossTier } from '../types';
+import type { Boss, BossCadence, BossDifficulty, BossTier } from '../types';
 import { getBossByFamily } from './bosses';
-import { hardestDifficulty, makeKey, parseKey, toggleBoss } from './bossSelection';
 
 /**
  * Boss Preset shortcuts for the Matrix Toolbar. Each preset is a fixed list
@@ -18,7 +17,67 @@ import { hardestDifficulty, makeKey, parseKey, toggleBoss } from './bossSelectio
  * the currently active preset deselects it. The drawer-level handler owns
  * this policy; the helpers below (`applyPreset` / `removePreset` /
  * `isPresetActive`) stay single-purpose and family-scoped.
+ *
+ * The selection-key grammar itself lives module-private inside
+ * `muleBossSlate.ts`; the two tiny helpers below duplicate only the string
+ * shape, because every key this module produces flows through
+ * `MuleBossSlate.from` or `slate.toggle` downstream where the Selection
+ * Invariant is actually enforced.
  */
+
+/** Build a native `<uuid>:<tier>:<cadence>` selection key. */
+function buildSelectionKey(
+  bossId: string,
+  tier: BossTier,
+  cadence: BossCadence,
+): string {
+  return `${bossId}:${tier}:${cadence}`;
+}
+
+/**
+ * Extract the bossId prefix from a selection key. Returns `null` for any
+ * string that doesn't have at least two colons, so malformed keys can pass
+ * through `removePreset` untouched.
+ */
+function keyBossIdPrefix(key: string): string | null {
+  const lastColon = key.lastIndexOf(':');
+  if (lastColon < 0) return null;
+  const tierColon = key.lastIndexOf(':', lastColon - 1);
+  if (tierColon < 0) return null;
+  return key.slice(0, tierColon);
+}
+
+/** Extract the `<cadence>` tail segment of a selection key, if present. */
+function keyCadenceSuffix(key: string): string | null {
+  const lastColon = key.lastIndexOf(':');
+  if (lastColon < 0) return null;
+  return key.slice(lastColon + 1);
+}
+
+/**
+ * Return the difficulty entry with the highest crystalValue for this boss.
+ * "Hardest" means "biggest numeric reward", irrespective of tier name or
+ * cadence — e.g. Vellum's weekly chaos beats its daily normal.
+ */
+function pickHardest(boss: Boss): BossDifficulty {
+  return boss.difficulty.reduce((best, d) =>
+    d.crystalValue > best.crystalValue ? d : best,
+  );
+}
+
+/** Resolve a preset entry to `{ boss, diff }`, or `null` on an unknown family/tier. */
+function resolveEntry(
+  entry: PresetFamily,
+): { boss: Boss; diff: BossDifficulty } | null {
+  const { family, tier } = entryInfo(entry);
+  const boss = getBossByFamily(family);
+  if (!boss) return null;
+  const diff = tier
+    ? boss.difficulty.find((d) => d.tier === tier)
+    : pickHardest(boss);
+  if (!diff) return null;
+  return { boss, diff };
+}
 
 export type PresetKey = 'CRA' | 'LOMIEN' | 'CTENE';
 
@@ -87,14 +146,9 @@ function entryInfo(entry: PresetFamily): { family: string; tier?: BossTier } {
  * family is unknown or the pinned tier isn't offered by the boss.
  */
 export function presetEntryKey(entry: PresetFamily): string | null {
-  const { family, tier } = entryInfo(entry);
-  const boss = getBossByFamily(family);
-  if (!boss) return null;
-  const diff = tier
-    ? boss.difficulty.find((d) => d.tier === tier)
-    : hardestDifficulty(boss);
-  if (!diff) return null;
-  return makeKey(boss.id, diff.tier, diff.cadence);
+  const resolved = resolveEntry(entry);
+  if (!resolved) return null;
+  return buildSelectionKey(resolved.boss.id, resolved.diff.tier, resolved.diff.cadence);
 }
 
 /** Family slug for a preset entry (unwraps the `{ family, tier }` form). */
@@ -103,9 +157,9 @@ export function presetEntryFamily(entry: PresetFamily): string {
 }
 
 /**
- * Swap each entry's target selection key into `keys`, using `toggleBoss`
- * semantics so any prior same-cadence sibling on that boss is replaced (and
- * opposite-cadence selections are preserved).
+ * Swap each entry's target selection key into `keys`. Any prior same-cadence
+ * sibling on that boss is replaced in-place (opposite-cadence selections are
+ * preserved so a mule keeps its daily + weekly selections side-by-side).
  */
 export function applyPreset(
   keys: string[],
@@ -113,16 +167,22 @@ export function applyPreset(
 ): string[] {
   let next = keys;
   for (const entry of families) {
-    const { family, tier } = entryInfo(entry);
-    const boss = getBossByFamily(family);
-    if (!boss) continue;
-    const diff = tier
-      ? boss.difficulty.find((d) => d.tier === tier)
-      : hardestDifficulty(boss);
-    if (!diff) continue;
-    const target = makeKey(boss.id, diff.tier, diff.cadence);
+    const resolved = resolveEntry(entry);
+    if (!resolved) continue;
+    const { boss, diff } = resolved;
+    const target = buildSelectionKey(boss.id, diff.tier, diff.cadence);
     if (next.includes(target)) continue; // idempotent
-    next = toggleBoss(next, boss.id, diff.tier);
+
+    // Same-cadence sibling on this boss: replace in-place so array order
+    // (and opposite-cadence selections) are preserved.
+    const siblingIdx = next.findIndex(
+      (k) => keyBossIdPrefix(k) === boss.id && keyCadenceSuffix(k) === diff.cadence,
+    );
+    if (siblingIdx >= 0) {
+      next = next.map((k, i) => (i === siblingIdx ? target : k));
+    } else {
+      next = [...next, target];
+    }
   }
   return next;
 }
@@ -143,9 +203,9 @@ export function removePreset(
     if (boss) dropFamilyIds.add(boss.id);
   }
   return keys.filter((k) => {
-    const parsed = parseKey(k);
-    if (!parsed) return true;
-    return !dropFamilyIds.has(parsed.bossId);
+    const bossId = keyBossIdPrefix(k);
+    if (bossId === null) return true;
+    return !dropFamilyIds.has(bossId);
   });
 }
