+++ src/modules/__tests__/income-toggle-integration.test.tsx
@@ -3,12 +3,11 @@ import { render, screen, fireEvent } from '@testing-library/react'
 import { IncomeProvider } from '../IncomeProvider'
 import { useFormatPreference, useMuleIncome, useTotalIncome } from '../income-hooks'
 import { bosses } from '../../data/bosses'
-import { makeKey } from '../../data/bossSelection'
 
 const LUCID = bosses.find((b) => b.family === 'lucid')!.id
 const WILL = bosses.find((b) => b.family === 'will')!.id
-const HARD_LUCID = makeKey(LUCID, 'hard', 'weekly')
-const HARD_WILL = makeKey(WILL, 'hard', 'weekly')
+const HARD_LUCID = `${LUCID}:hard:weekly`
+const HARD_WILL = `${WILL}:hard:weekly`
 
 function MuleIncomeDisplay({ mule, index }: { mule: { selectedBosses: string[] }; index: number }) {
   const income = useMuleIncome(mule)
