+++ src/modules/income.ts
@@ -1,5 +1,4 @@
-import { getBossById } from '../data/bosses'
-import { parseKey } from '../data/bossSelection'
+import { MuleBossSlate } from '../data/muleBossSlate'
 import { formatMeso } from '../utils/meso'
 
 export interface IncomeDisplay {
@@ -8,25 +7,15 @@ export interface IncomeDisplay {
 }
 
 /**
- * Sum crystalValues for a list of `<uuid>:<tier>` selection keys, folding
- * daily tiers into the weekly headline at `crystalValue × 7`.
+ * Sum crystalValues for a list of `<uuid>:<tier>:<cadence>` selection keys,
+ * folding daily tiers into the weekly headline at `crystalValue × 7`.
  *
- * The source of truth for cadence is `BossDifficulty.cadence` — daily tiers
- * are farmable up to 7× per week, weekly tiers clear once. Unknown bosses
- * and tiers no longer offered contribute 0.
+ * Goes through `MuleBossSlate.from` so unresolvable keys and
+ * duplicate-bucket losers drop silently — the same normalization any
+ * persisted slate would receive.
  */
 export function sumSelectedKeys(keys: string[]): number {
-  let total = 0
-  for (const key of keys) {
-    const parsed = parseKey(key)
-    if (!parsed) continue
-    const boss = getBossById(parsed.bossId)
-    if (!boss) continue
-    const diff = boss.difficulty.find((d) => d.tier === parsed.tier)
-    if (!diff) continue
-    total += diff.cadence === 'daily' ? diff.crystalValue * 7 : diff.crystalValue
-  }
-  return total
+  return MuleBossSlate.from(keys).totalCrystalValue
 }
 
 export function computeMuleIncome(selectedBosses: string[], abbreviated: boolean): IncomeDisplay {
